IS: introduce JS setting to know if there are enough entries to scroll at least once.
The ::is_last_batch method was rewritten to more reliable reflect if it's the last batch. There were cases where it failed, for example: when there were 15 posts and IS displayed 3 posts per page, the last ::is_last_batch compared 3 posts returned from db > 3 posts per page, which was false, and so IS thought there were most posts to fetch. It ran one more time but returned nothing, which is a confusing UX.
This commit also addresses an issue of a missing dependency: when there were 3 posts and IS displayed 3 posts per page, it didn't enqueued the assets in ::action_template_redirect but the theme compatibility still required it. They are now registered regardless of whether they're enqueued or not, which solves the dependency issue.